backend/requirements.txt
Django==3.2.15
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
Pillow==9.2.0